Make CLI able to run headless
